requirements.txt
langchain
langchain_community
streamlit
python_dotenv 